+++ riscv-rt/src/lib.rs
@@ -371,24 +371,6 @@
 //! );
 //! ```
 //!
-//! ## `_setup_interrupts`
-//!
-//! This function is called right before the main function and is responsible for setting up
-//! the interrupt controller.
-//!
-//! Default implementation sets the trap vector to `_start_trap` in direct mode.
-//! If the `v-trap` feature is enabled, the trap vector is set to `_vector_table`
-//! in vectored mode. Users can override this function by defining their own `_setup_interrupts`.
-//!
-//! This function can be redefined in the following way:
-//!
-//! ``` no_run
-//! #[export_name = "_setup_interrupts"]
-//! pub fn setup_interrupts() {
-//!    // ...
-//! }
-//! ```
-//!
 //! ## `hal_main`
 //!
 //! Internally, `riscv-rt` does not jump to the `main` function created by the user using the
@@ -561,6 +543,28 @@
 //!
 //! You can use the [`#[post_init]`][attr-post-init] attribute to define a post-init function with Rust.
 //!
+//! ## `custom-setup-interrupts`
+//!
+//! The `riscv-rt` crate provides a default implementation for the `_setup_interrupts` function.
+//! This function is called right before the main function and is responsible for setting up the interrupt controller.
+//! Default implementation sets the trap vector to `_start_trap` in direct mode. If the `v-trap` feature is enabled,
+//! the trap vector is set to `_vector_table` in vectored mode.
+//!
+//! However, in some cases, users may want to provide their own implementation of this function to customize the interrupt
+//! setup process. Users can override this function by:
+//!
+//! 1. Enabling the `custom-setup-interrupts` feature to opt-out the default implementation.
+//! 2. Using the [`#[setup_interrupts]`][attr-setup-interrupts] attribute on their custom function.
+//!
+//! This function can be redefined in the following way:
+//!
+//! ``` no_run
+//! #[riscv_rt::setup_interrupts]
+//! fn setup_interrupts(hart_id: usize) {
+//!    // ...
+//! }
+//!```
+//!
 //! ## `single-hart`
 //!
 //! Saves a little code size if there is only one hart on the target.
@@ -617,9 +621,10 @@
 //!
 //! ## `u-boot`
 //!
-//! When the U-boot feature is enabled, acceptable signature for `#[entry]` macros is changed. This is required
-//! because when booting from elf, U-boot passes `argc` and `argv`. This feature also implies `single-hart`.
-//! The only way to get boot-hart is through fdt, so other harts initialization is up to you.
+//! When the U-boot feature is enabled, acceptable signature for `#[post_init]`, `#[setup_interrupts]`,
+//! and`#[entry]` macros is changed. This is required because when booting from elf, U-boot passes
+//! `argc` and `argv`. This feature also implies `single-hart`. The only way to get boot-hart is through
+//! fdt, so other harts initialization is up to you.
 //!
 //! ## `pre-default-start-trap`
 //!
@@ -672,6 +677,7 @@
 //! [attr-external-interrupt]: attr.external_interrupt.html
 //! [attr-core-interrupt]: attr.core_interrupt.html
 //! [attr-post-init]: attr.post_init.html
+//! [attr-setup-interrupts]: attr.setup_interrupts.html
 
 // NOTE: Adapted from cortex-m/src/lib.rs
 #![no_std]
@@ -687,22 +693,23 @@ pub mod exceptions;
 pub mod interrupts;
 
 #[cfg(feature = "s-mode")]
-use riscv::register::{
-    scause as xcause,
-    stvec::{self as xtvec, Stvec as Xtvec, TrapMode},
-};
+use riscv::register::scause as xcause;
+#[cfg(all(feature = "s-mode", not(feature = "custom-setup-interrupts")))]
+use riscv::register::stvec::{self as xtvec, Stvec as Xtvec, TrapMode};
 
 #[cfg(not(feature = "s-mode"))]
-use riscv::register::{
-    mcause as xcause,
-    mtvec::{self as xtvec, Mtvec as Xtvec, TrapMode},
-};
+use riscv::register::mcause as xcause;
+#[cfg(all(not(feature = "s-mode"), not(feature = "custom-setup-interrupts")))]
+use riscv::register::mtvec::{self as xtvec, Mtvec as Xtvec, TrapMode};
 
 pub use riscv_macros::{core_interrupt, entry, exception, external_interrupt};
 pub use riscv_types::*;
 
 #[cfg(feature = "post-init")]
 pub use riscv_macros::post_init;
+
+#[cfg(feature = "custom-setup-interrupts")]
+pub use riscv_macros::setup_interrupts;
 
 /// We export this static with an informative name so that if an application attempts to link
 /// two copies of riscv-rt together, linking will fail. We also declare a links key in
@@ -729,13 +736,13 @@ pub unsafe extern "C" fn start_rust(a0: usize, a1: usize, a2: usize) -> ! {
     extern "Rust" {
         #[cfg(feature = "post-init")]
         fn __post_init(a0: usize);
-        fn _setup_interrupts();
+        fn _setup_interrupts(a0: usize);
         fn hal_main(a0: usize, a1: usize, a2: usize) -> !;
     }
 
     #[cfg(feature = "post-init")]
     __post_init(a0);
-    _setup_interrupts();
+    _setup_interrupts(a0);
     hal_main(a0, a1, a2);
 }
 
@@ -746,14 +753,18 @@ pub unsafe extern "C" fn start_rust(a0: usize, a1: usize, a2: usize) -> ! {
 ///
 /// # Note
 ///
-/// Users can override this function by defining their own `_setup_interrupts` function.
+/// You can define your own `_setup_interrupts` function to override the default implementation.
+/// To do so, you must enable the `custom-setup-interrupts` feature to opt-out the default implementation.
+/// Then, you can use the [`riscv_macros::setup_interrupts`] attribute on your custom function.
+/// This macro is re-exported by this crate if the `custom-setup-interrupts` feature is enabled.
 ///
 /// # Safety
 ///
 /// This function should not be called directly by the user, and should instead
 /// be invoked by the runtime implicitly. It is expected to be called before the main function.
-#[export_name = "_default_setup_interrupts"]
-pub unsafe extern "Rust" fn setup_interrupts() {
+#[cfg(not(feature = "custom-setup-interrupts"))]
+#[riscv_macros::setup_interrupts]
+unsafe fn default_setup_interrupts() {
     extern "C" {
         #[cfg(not(feature = "v-trap"))]
         fn _start_trap();