+++ tests-trybuild/tests/riscv-rt/setup_interrupts/pass_empty_unsafe.rs
@@ -1,0 +1,4 @@
+#[riscv_macros::setup_interrupts]
+unsafe fn setup_interrupts() {}
+
+fn main() {}